refactor(server): extract per-terminal provider DAG into providers.ts
Carves the seven per-terminal observers (agent-command tracker, git
watcher, GitHub PR watcher, foreground-process observer, three agent
detectors) out of `LocalTerminalBackend` and into a new
`packages/server/src/terminalBackend/providers.ts`. The DAG is now
parameterized over `ProviderHooks` + `ProviderChannels` + `ProviderRecord`
so any future host backend (a remote `kolu --stdio` agent in #951 R-2,
say) can run the same implementation against its own channels and
metadata-publish surface — no copy/paste, single source of truth.

`local.ts` shrinks from 763 → 332 LoC. The semantics are byte-identical:
each `spawnPty` constructs the record, builds per-terminal channels
mapped onto `terminalChannels.{cwd,title,commandRun,git}(id)`, and
wires hooks that funnel writes through `updateServer*Metadata` and
`trackRecent*`. Provider ordering, the lazy external-changes activation,
the `setAgentMetadata` two-call shape (live + recency bump), and the
`COMMAND_RUN_RECONCILE_DELAYS_MS` retry schedule all preserved verbatim.

README §Architecture and `packages/surface/README.md` updated to point
provider-DAG references at the new file.

Part of #951.
